tab completion for targets working. Added function to cache targets, added completion function to main to tab complete targets
