+++ yotta/target.py
@@ -33,28 +33,18 @@ Target_RE = re.compile('^('+
     ')?'+
 ')$')
 
-def targetCompletion(prefix, parsed_args, **kwargs):
-    t = cache.get('targets')
-    logging.debug("Target List = %s" %t)
-    print("Trying To Tab Complete... should return the following...")
-    for thing in t:
-        if thing.startswith(prefix):
-            print(thing)
-    return(x for x in t if x.startswith(prefix))
-
 def addOptions(parser):
     parser.add_argument('set_target', default=None, nargs='?',
         help='set the build target to this (targetname[,versionspec_or_url])'
-    ).ChoicesCompleter = targetCompletion
+    )
     parser.add_argument('-g', '--global', dest='save_global',
         default=True, action='store_true',
         help='set globally (in the per-user settings) instead of locally to this directory'
     )
-    parser.add_argument('-u','--updatelocal', 
-        dest='updateLocal', 
-        default=False, action='store_true',
+    parser.add_argument('-u','--updatelocal',
+        dest='updateLocal', default=True,action='store_true',
         help='update the local cache of targets available for tab completion'
-    ).completer = targetCompletion
+    )
 
     # FIXME: need help that lists possible targets, and we need a walkthrough
     # guide to forking a new target for an existing board
@@ -125,13 +115,11 @@ def displayCurrentTarget(args):
 
 
 def execCommand(args, following_args):
+    if args.updateLocal is True:
+        updateTargetList()
     if args.set_target is None:
         return displayCurrentTarget(args)
     else:
-        # TODO: update code so it searches all registries saved to system. (private, public, and local)
-        if args.updateLocal is True:
-            updateTargetList()
-            targetCompletion("frdm",0)            
         if not Target_RE.match(args.set_target):
             logging.error('''Invalid target: "%s"''' % args.set_target)#, targets must be one of:
             #
